requirements.txt
streamlit
PyPDF2
langchain
langchain_google_genai
google-generativeai
faiss-cpu
python-dotenv
